Updated the code to retrieve the user rating
The output text file consists of the individual user reviews along with the rating of each comment for a given movie. 
As of now this code runs only for one movie. If we want the code to run for all the movies in the imdb dataset, we can uncomment the line number 63 and comment the line 64.